feat(post): adjusted component + implemented delete post + settings
